+++ main.py
@@ -500,7 +500,7 @@ if __name__ == "__main__":
     app.setApplicationName("AIPacs")
     # app.setApplicationDisplayName("AIPacs - Professional Medical Imaging Suite")
     app.setApplicationDisplayName("AIPacs")
-    app.setApplicationVersion("2.4.4")
+    app.setApplicationVersion("2.4.5")
     app.setOrganizationName("AIPacs")
 
     # Setup font rendering for better quality